third_party/WebKit/LayoutTests/virtual/threaded/fast/animationworklet/animation-worklet-animator-animate.html

<!DOCTYPE html>

<script id="simple_animate" type="text/worklet">
registerAnimator("test_animator", class {
  constructor() {
    console.log("test animator constructor called.");
  }
  animate(currentTime, effect) {
    if (!this.animated)
        console.log("test animator animate called.");

    this.animated = true;
  }
});
</script>

<div id="target" style="will-change: transform, opacity;"></div>

<script src="resources/animation-worklet-tests.js"></script>
<script>
if (window.testRunner) {
  testRunner.waitUntilDone();
  testRunner.dumpAsText();
}

runInAnimationWorklet(
  document.getElementById('simple_animate').textContent
).then(_ => {
  const effect = new KeyframeEffect(document.getElementById("target"), [{ opacity: 0 }], { duration: 1000 });
  const animation = new WorkletAnimation('test_animator', [effect], document.timeline, {});
  animation.play();

  if (window.testRunner) {
    waitTwoAnimationFrames( _ => {
      testRunner.notifyDone();
    });
  }
});
</script>
