third_party/WebKit/LayoutTests/virtual/threaded/fast/animationworklet/animation-worklet-animator-animate-expected.txt

CONSOLE MESSAGE: line 4: test animator constructor called.
CONSOLE MESSAGE: line 8: test animator animate called.
